requirements.txt
torch
diffusers>=0.30.0
transformers>=4.40.0
accelerate
gradio>=4.0.0
Pillow
sentencepiece
protobuf
safetensors
